🚚 Migrate `about` to app directory
