+++ src/app/about/components/Experience/Job/CompanyLogo/index.tsx
