+++ src/app/about/components/Travelling/Map/index.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 const Map = () => (
   <>
     <style jsx global>
